refactor(admin-ui): Phase2-P1-2 PR2 — ApiKeysTab・EmbedCodeTab を別ファイルに分離 + 共有 types.ts 導入 (GID 1214120401966111)
[id].tsx を 2911→2426行に削減。機能変更なし。
- types.ts (48行): TenantFeatures・TenantDetail・ApiKey・CARD_STYLE を共有エクスポート
- ApiKeysTab.tsx (289行): APIキー管理 UI + fetchApiKeys/revokeApiKey helper を独立ファイルに抽出
- EmbedCodeTab.tsx (186行): 埋め込みコード生成 UI を独立ファイルに抽出
- [id].tsx: 上記インターフェース・スタイル定数・関数を types.ts/ApiKeysTab.tsx から import に変更
